+++ tests/test_asset_zoo.py
@@ -1,7 +1,11 @@
 import mujoco
 import pytest
 
-from mjlab.asset_zoo.robots import get_g1_robot_cfg, get_go1_robot_cfg
+from mjlab.asset_zoo.robots import (
+  get_g1_robot_cfg,
+  get_g1_with_hands_robot_cfg,
+  get_go1_robot_cfg,
+)
 from mjlab.entity import Entity
 
 
@@ -9,6 +13,7 @@ from mjlab.entity import Entity
   "robot_name,robot_cfg_fn",
   [
     ("G1", get_g1_robot_cfg),
+    ("G1_WITH_HANDS", get_g1_with_hands_robot_cfg),
     ("GO1", get_go1_robot_cfg),
   ],
 )